requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-redis==5.4.0
python-decouple==3.8
Pillow==10.0.1
django-extensions==3.2.3
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
gunicorn==21.2.0
whitenoise==6.6.0